% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bakeLasagna.R
\name{bakeLasagna}
\alias{bakeLasagna}
\title{Create side-by-side lasagna plot and density plot}
\usage{
bakeLasagna(data, data.long, covariate = NULL)
}
\arguments{
\item{data}{Dataset for lasagna plot. Same data used in \code{makeLasagna()} function.}

\item{data.long}{Sorted longform dataset for lasagna plot output by \code{makeLasagna()} function.}

\item{covariate}{User-selected covariate for sorting the rows in the lasagna plot. 
Defaults to NULL, in which case data is sorted by row number.}
}
\description{
Internal method used in conjunction with \code{makeLasagna()} to create side-by-side lasagna plot and distribution plot. 
The distribution plot gives distribution of sorting covariate.
}
\author{
Julia Wrobel \email{jw3134@cumc.columbia.edu}

Nicole Marie Lapointe Jameson
}
