% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/detectOutliers.R
\name{combinat}
\alias{combinat}
\title{internal function from 'fda' package}
\usage{
combinat(n, p)
}
\arguments{
\item{n}{number of columns in your dataset}

\item{p}{number of rows in your dataset}
}
\description{
function used in method for fast modified band depth (MBD) calculation
}
\author{
Ying Sun and Marc G.Genton
}

