% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createInputCall.R
\name{createInputCall}
\alias{createInputCall}
\title{Create input calls for plot_shiny.fosr()}
\usage{
createInputCall(name, variable)
}
\arguments{
\item{name}{variable name}

\item{variable}{variable values from dataset}
}
\description{
Internal method that constructs the input calls for plot_shiny.fosr(). The
variable name and values are passed as arguments, and a corresponding slider (for
numeric) or drop-down (for factor) input is constructed.
}
\author{
Jeff Goldsmith \email{ajg2202@cumc.columbia.edu}
}

