% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createInvLink.R
\name{createInvLink}
\alias{createInvLink}
\title{Return inverse link function for plot_shiny.fpca()}
\usage{
createInvLink(family = NULL)
}
\arguments{
\item{family}{Family of the (generalized) FPCA. Currently supported families
are \code{gaussian} and \code{binomial}.}
}
\description{
Internal method that constructs the inverse link function for a 
generalized FPCA fit. This is used in toggling between plots on the
natural scale and on the response scale.
}
\author{
Jeff Goldsmith \email{ajg2202@cumc.columbia.edu}
}

