% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadModule.R
\name{downloadModule}
\alias{downloadModule}
\title{download Plot as PDF or ggplot Object, modularized server}
\usage{
downloadModule(input, output, session, plotObject, plotName)
}
\arguments{
\item{input}{gets user input from UI}

\item{output}{designates output for UI}

\item{session}{Shiny variable for server modules}

\item{plotObject}{Reactive plot object defined elsewhere in the server function.}

\item{plotName}{Character string designating name of the plot for PDF output.}
}
\description{
Internal method that creates UI with buttons to download a plot as a PDF or ggplot object.
}
\author{
Julia Wrobel \email{jw3134@cumc.columbia.edu}
}

