% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadModule.R
\name{downloadModuleUI}
\alias{downloadModuleUI}
\title{download Plot as PDF or ggplot Object, modularized UI}
\usage{
downloadModuleUI(id)
}
\arguments{
\item{id}{name of module. Allows each call of this module to be uniquely identified.}
}
\description{
Internal method that creates UI with buttons to download a plot as a PDF or ggplot object.
}
\author{
Julia Wrobel \email{jw3134@cumc.columbia.edu}
}

