% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registerLasagna.R
\name{getWidth}
\alias{getWidth}
\title{Get spaces between timepoints as widths for binary registration lasagna plot.}
\usage{
getWidth(z)
}
\arguments{
\item{z}{time values for a specific subject}
}
\description{
Get spaces between timepoints as widths for binary registration lasagna plot.
}
\author{
Julia Wrobel \email{jw3134@cumc.columbia.edu}
}
