% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeDF.R
\name{makeDF}
\alias{makeDF}
\title{make melted dataframes for mfpca}
\usage{
makeDF(IDvalue, data, subsetON, ...)
}
\arguments{
\item{IDvalue}{id on which to create dataframe.}

\item{data}{name of data to be manipulated.}

\item{subsetOn}{name of dataframe with 'id' column on which to subset data.}
}
\value{
a dataframe of melted observations or Yhat values for use in ggplot.
}
\description{
Function to create dataframes of observed and fitted values for plotting with mfpca.
}
\author{
Julia Wrobel
}

