% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLasagna.R
\name{makeLasagna}
\alias{makeLasagna}
\title{Pre-process data for lasagna plot}
\usage{
makeLasagna(data = DTI, outcome, covariate = NULL)
}
\arguments{
\item{data}{Dataset for lasagna plot.}

\item{outcome}{Matrix of values where each row represents a functional observation.}

\item{covariate}{User-selected covariate for sorting the rows in the lasagna plot. 
Defaults to NULL, in which case data is sorted by row number.}
}
\description{
Internal method that takes a dataframe of observed data with an outcome matrix and user-selected covariate, 
sorts outcome by the selected covariate, and assigns heights to each row based on value of the selected covariate. 
The resulting dataframe is used with \code{bakeLasagna()} to create lasagna plot.
}
\author{
Julia Wrobel \email{ajg2202@cumc.columbia.edu}

Nicole Marie Lapointe Jameson \email{nml2145@columbia.edu}
}

