% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/createInputCall.R
\name{mfpcaCalls}
\alias{mfpcaCalls}
\title{Create input calls for plot_shiny.mfpca()}
\usage{
mfpcaCalls(plot.npc, plotObj, percents)
}
\arguments{
\item{plot.npc}{list of 2 numeric entries giving number of sliders at each level}

\item{plotObj}{the mfpca object plotted in the plot_shiny.mfpca() function.}

\item{percents}{the percent variance calculated for each eigen values for levels 1 and 2.}
}
\value{
a list of numbers that indicate percent variance for selected level.
}
\description{
Internal method that constructs the input calls for plot_shiny.mfpca(). The
number of sliders to construct for each level is passed as an argument, and
corresponding sliders for each FPC are constructed.
}
\author{
Julia Wrobel \email{jw3134@cumc.columbia.edu}
}

