% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/detectOutliers.R
\name{outliers}
\alias{outliers}
\title{Identifies outliers for plot_shiny.fosr()}
\usage{
outliers(data, factor = 1.5)
}
\arguments{
\item{data}{matrix or df of functional observations}

\item{factor}{a constant that determines the fences for outliers. Defaults to 1.5, as in classical definition for Tukey outliers.}
}
\description{
Internal method that assigns band depth values to curves based on exact fast MBD computation (Sun & Genton, 2012).
Code modified from fbplot in fda package.
A dataframe of residuals is passed as an argument, and depths and outlying curves are returned
}
\author{
Julia Wrobel \email{jw3134@cumc.columbia.edu}
}

