% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/plot.shiny.fosr.R
\name{plot.shiny.fosr}
\alias{plot.shiny.fosr}
\title{Interactive Plotting for Functional-on-Scalar Regressions}
\usage{
\method{plot}{shiny.fosr}(fosr.obj, xlab = "", ylab = "", title = "")
}
\arguments{
\item{fosr.obj}{fosr object to be plotted.}

\item{xlab}{x axis label}

\item{ylab}{y axis label}

\item{title}{plot title}
}
\description{
Produces an interactive plot illustrating a function-on-scalar
regression analysis.
}
\author{
Jeff Goldsmith \email{jeff.goldsmith@columbia.edu},
Julia Wrobel \email{jw3134@cumc.columbia.edu}
}
\seealso{
\code{\link{plot.shiny}}
}

