% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/plot.shiny.fpca.R
\name{plot.shiny.fpca}
\alias{plot.shiny.fpca}
\title{Interactive Plotting for Functional Principal Component Analysis}
\usage{
\method{plot}{shiny.fpca}(fpca.obj, xlab = "", ylab = "", title = "")
}
\arguments{
\item{fpca.obj}{fpca object to be plotted.}

\item{xlab}{x axis label}

\item{ylab}{y axis label}

\item{title}{plot title}
}
\description{
Produces an interactive plot illustrating a functional principal component
analysis.
}
\author{
Julia Wrobel \email{jw3134@cumc.columbia.edu},
Jeff Goldsmith \email{jeff.goldsmith@columbia.edu}
}
\seealso{
\code{\link{plot.shiny}}
}

