% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shiny.flcm.R
\name{plot_shiny.flcm}
\alias{plot_shiny.flcm}
\title{Interactive Plotting for Functional Linear Concurrent regression}
\usage{
\method{plot_shiny}{flcm}(obj, xlab = "", ylab = "", title = "", ...)
}
\arguments{
\item{obj}{fosr object to be plotted.}

\item{xlab}{x axis label}

\item{ylab}{y axis label}

\item{title}{plot title}

\item{...}{additional arguments passed to plotting functions}
}
\description{
Produces an interactive plot illustrating a functional linear concurrent
regression analysis.
}
\seealso{
\code{\link{plot_shiny}}
}
\author{
Jeff Goldsmith \email{jeff.goldsmith@columbia.edu},
Julia Wrobel \email{jw3134@cumc.columbia.edu}
}
