% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shiny.fpca.R
\name{plot_shiny.fpca}
\alias{plot_shiny.fpca}
\title{Interactive Plotting for Functional Principal Component Analysis}
\usage{
\method{plot_shiny}{fpca}(obj, xlab = "", ylab = "", title = "", ...)
}
\arguments{
\item{obj}{fpca object to be plotted.}

\item{xlab}{x axis label}

\item{ylab}{y axis label}

\item{title}{plot title}

\item{...}{additional arguments passed to plotting functions}
}
\description{
Produces an interactive plot illustrating a functional principal component
analysis.
}
\seealso{
\code{\link{plot_shiny}}
}
\author{
Julia Wrobel \email{jw3134@cumc.columbia.edu},
Jeff Goldsmith \email{jeff.goldsmith@columbia.edu}
}
