% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_shiny.lfpca.R
\name{plot_shiny.lfpca}
\alias{plot_shiny.lfpca}
\title{Interactive Plotting for Longtiudinal Functional Data Analysis using FPCA}
\usage{
\method{plot_shiny}{lfpca}(obj, xlab = "", ylab = "", title = "", ...)
}
\arguments{
\item{obj}{lfpca object to be plotted.}

\item{xlab}{x axis label}

\item{ylab}{y axis label}

\item{title}{plot title}

\item{...}{additional arguments passed to plotting functions}
}
\description{
Produces an interactive plot illustrating longitudinal functioanl data analysis (Park and Staicu, 2015).
}
\author{
So Young Park \email{spark13@ncsu.edu}, Ana-Maria Staicu \email{astaicu@ncsu.edu}
}
\references{
Park, S.Y. and Staicu, A.M. (2015). Longitudinal functional data analysis. Stat 4 212-226.
}
\seealso{
\code{\link{plot_shiny}}; \code{fpca.lfda} in the refund package for estimation method.
}

