% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shiny.registration.R
\name{plot_shiny.registration}
\alias{plot_shiny.registration}
\title{Interactive Plotting for Registration Objects}
\usage{
\method{plot_shiny}{registration}(obj, xlab = "", ylab = "", title = "",
  ...)
}
\arguments{
\item{obj}{registration object to be plotted.}

\item{xlab}{x axis label}

\item{ylab}{y axis label}

\item{title}{plot title}

\item{...}{additional arguments passed to plotting functions}
}
\description{
Produces an interactive plot illustrating functional data before and after registration.
Our registration method uses FPCA, the FPCA is plotted as well.
}
\seealso{
\code{\link{plot_shiny}}
}
\author{
Julia Wrobel \email{jw3134@cumc.columbia.edu}
}
