% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registerLasagna.R
\name{registerLasagna}
\alias{registerLasagna}
\title{Create lasagna plot for unregistered and registered data}
\usage{
registerLasagna(data)
}
\arguments{
\item{data}{Dataset for lasagna plot.}
}
\description{
Get registered and unregistered lasagna plots for binary data. Note: should make this compatible
for other data types as well. Requires data to have t_hat and tstar variables.
}
\author{
Julia Wrobel \email{jw3134@cumc.columbia.edu}
}
