% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/downloadPlot.R
\name{savePDF}
\alias{savePDF}
\title{Save Plot Object as PDF}
\usage{
savePDF(title, plotName)
}
\arguments{
\item{title}{new name for the plot, and name of the PDF file created}

\item{plotName}{name of the ggplot object}
}
\description{
Internal method that saves plots as PDF.Can be used with all plotting methods in the package. The
name of the plot object and it's name to be saved under are passed in and the plot is saved as a PDF.
}
\author{
Julia Wrobel \email{jw3134@cumc.columbia.edu}
}

