% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadPlot.R
\name{savePlot}
\alias{savePlot}
\title{Save Plot Object as .RData file}
\usage{
savePlot(title, plotName)
}
\arguments{
\item{title}{new name for the plot, and name of the RData file created.}

\item{plotName}{name of the ggplot object.}
}
\description{
Internal method that saves ggplot plots as .RData files.Can be used with all plotting methods in the package. The
name of the plot object and it's name to be saved under are passed in and the plot is saved as an RData file.
}
\author{
Julia Wrobel \email{jw3134@cumc.columbia.edu}
}
