% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thin_functional_data.R
\name{thin_functional_data}
\alias{thin_functional_data}
\title{Thin functional data}
\usage{
thin_functional_data(Y, length_out = 100)
}
\arguments{
\item{Y}{functional dataframe}

\item{length_out}{number of points per subject for dataframe that is returned}
}
\description{
Takes a dense functional dataset in long form and thins it so that there are 100 observations per
subject, equally spaced.
}
\author{
Julia Wrobel \email{jw3134@cumc.columbia.edu}
}
