% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/varPercent.R
\name{varPercent}
\alias{varPercent}
\title{Calculate percent variance of eigenvalues for plot_shiny.mfpca()}
\usage{
varPercent(level, plotObj)
}
\arguments{
\item{level}{numeric, 1 or 2 for levels 1 or 2, respectively, 12 to calculate total variance.}

\item{plotObj}{the mfpca object plotted in the plot_shiny.mfpca() function.}
}
\value{
a list of numbers that indicate percent variance for selected level.
}
\description{
Internal method that calculates percent variance of eigenvalues for specified level (1, 2, or total) for
plot_shiny.mfpca(). The desired level is passed in as an argument (level = 12 for total) and a list of percent
variances is returned.
}
\author{
Julia Wrobel \email{jw3134@cumc.columbia.edu}
}

