% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/detectOutliers.R
\name{combinat}
\alias{combinat}
\title{function used in method for fast modified band depth (MBD) calculation}
\usage{
combinat(n, p)
}
\arguments{
\item{n}{}

\item{p}{}
}
\description{
function used in method for fast modified band depth (MBD) calculation
}
\author{
Ying Sun and Marc G.Genton
}

