% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/detectOutliers.R
\name{fMBD}
\alias{fMBD}
\title{fast modified band depth calculation for fda}
\usage{
fMBD(data)
}
\arguments{
\item{data}{name of dataset}
}
\description{
Method for fast modified band depth (fMBD) calculation
}
\author{
Ying Sun and Marc G.Genton
}

