% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_muPC.R
\name{make_muPC}
\alias{make_muPC}
\title{Create muPC plot for FPCA panels}
\usage{
make_muPC(obj, pc_choice, response_scale = FALSE)
}
\arguments{
\item{obj}{fpca object to be plotted.}

\item{pc_choice}{FPC to be plotted.}

\item{response_scale}{Scale of reponse to be plotted. If TRUE results are plotted on response scale,
if FALSE results are plotted on natural scale.}
}
\description{
Produces a ggplot with mean plus or minus two standard deviations of a selected FPC.
}
