% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/createInputCall.R
\name{mfpcaCalls}
\alias{mfpcaCalls}
\title{Create input calls for plot_shiny.mfpca()}
\usage{
mfpcaCalls(numSliders = NULL, plotObj, percents)
}
\arguments{
\item{numSliders}{numeric, number of PCs per level to create sliders for. By default it uses number of PCs.}

\item{plotObj}{the mfpca object plotted in the plot_shiny.mfpca() function.}

\item{percents}{the percent variance calculated for each eigen values for levels 1 and 2.}
}
\value{
a list of numbers that indicate percent variance for selected level.
}
\description{
Internal method that constructs the input calls for plot_shiny.mfpca(). The
number of sliders to construct for each level is passed as an argument, and
corresponding sliders for each FPC are constructed.
}
\author{
Julia Wrobel \email{jw3134@cumc.columbia.edu}
}

