% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectOutliers.R
\name{outliers}
\alias{outliers}
\title{Identifies outliers for plot_shiny.fosr()}
\usage{
outliers(data, factor = 1.5)
}
\arguments{
\item{data}{matrix or df of functional observations}

\item{factor}{a constant that determines the fences for outliers. Defaults to 1.5, as in classical definition for Tukey outliers.}
}
\description{
Internal method that assigns band depth values to curves based on exact fast MBD computation (Sun & Genton, 2012).
Code modified from fbplot in fda package. 
A dataframe of residuals is passed as an argument, and depths and outlying curves are returned
}
\references{
Sun, Ying, Marc G. Genton, and Douglas W. Nychka. (2012).
Exact fast computation of band depth for large functional datasets: How quickly can one million curves be ranked? \emph{Stat}, 1, 68-74.

Sun, Ying, and Marc G. Genton. (2011). Functional boxplots. \emph{Journal of Computational and Graphical Statistics}, 20, 313-334.
}
\author{
Julia Wrobel \email{jw3134@cumc.columbia.edu}
}
