% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shiny.fosr.R
\name{plot_shiny.fosr}
\alias{plot_shiny.fosr}
\title{Interactive Plotting for Functional-on-Scalar Regressions}
\usage{
\method{plot_shiny}{fosr}(obj, xlab = "", ylab = "", title = "", ...)
}
\arguments{
\item{obj}{fosr object to be plotted.}

\item{xlab}{x axis label}

\item{ylab}{y axis label}

\item{title}{plot title}

\item{...}{additional arguments passed to plotting functions}
}
\description{
Produces an interactive plot illustrating a function-on-scalar
regression analysis.
}
\author{
Jeff Goldsmith \email{jeff.goldsmith@columbia.edu},
Julia Wrobel \email{jw3134@cumc.columbia.edu}
}
\seealso{
\code{\link{plot_shiny}}
}

