% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shiny.fpca.R
\name{plot_shiny.fpca}
\alias{plot_shiny.fpca}
\title{Interactive Plotting for Functional Principal Component Analysis}
\usage{
\method{plot_shiny}{fpca}(obj, xlab = "", ylab = "", title = "", thin_data = FALSE, ...)
}
\arguments{
\item{obj}{fpca object to be plotted.}

\item{xlab}{x axis label}

\item{ylab}{y axis label}

\item{title}{plot title}

\item{thin_data}{If TRUE data is thinned for each subject to make plotting faster. Defaults to FALSE.}

\item{...}{additional arguments passed to plotting functions}
}
\description{
Produces an interactive plot illustrating a functional principal component
analysis.
}
\seealso{
\code{\link{plot_shiny}}
}
\author{
Julia Wrobel \email{julia.wrobel@cuanschutz.edu},
Jeff Goldsmith \email{jeff.goldsmith@columbia.edu}
}
