% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shiny.mfpca.R
\name{plot_shiny.mfpca}
\alias{plot_shiny.mfpca}
\title{Interactive Plotting for Multilevel Functional Principal Component Analysis}
\usage{
\method{plot_shiny}{mfpca}(obj, xlab = "", ylab = "", title = "", ...)
}
\arguments{
\item{obj}{mfpca object to be plotted.}

\item{xlab}{x axis label}

\item{ylab}{y axis label}

\item{title}{plot title}

\item{...}{additional arguments passed to plotting functions}
}
\value{
No object is returned. This function takes in objects of class 'mfpca' and outputs a shiny application for that object.
}
\description{
Produces an interactive plot illustrating a multilevel functional principal component
analysis.
}
\seealso{
\code{\link{plot_shiny}}
}
\author{
Julia Wrobel \email{julia.wrobel@cuanschutz.edu},
Jeff Goldsmith \email{jeff.goldsmith@columbia.edu}
}
